5 approval, 5 days passed, Merged, Thanks @derrel!
* Subscriber list can be altered mid-stream.

The message bus subscriber list is iterated, and each subscribed function is
called in turn. If one of them unsubscribes itself, the subscriber list
decreased in length, and the loop failed to iterate over all original
subscribers. With this change, the subscriber list is cloned at the beginning,
and each subscribed function is called properly.

This fixes a real bug. A subscriber unsubscribing himself would cause later
subscribers in the list to never be called. This *may*, however, also be a
backward compatability breakage, in the rare case that an early subscriber
unsubscribes a later subscriber. Formerly, that later subscriber would not
have been called during this loop iteration. In the current implementation,
all subscribers originally in the list will be called.

* Fix documentation: was showing wrong parameter type